background video blocks now working
