+++ network-api/networkapi/wagtailpages/pagemodels/blog/blog.py
@@ -1,6 +1,5 @@
 from django.db import models
 from django.conf import settings
-from django.core.exceptions import ValidationError
 
 from wagtail.admin.edit_handlers import (
     FieldPanel,
@@ -14,7 +13,6 @@ from wagtail.core import blocks
 from wagtail.core.models import Orderable, Page
 from wagtail.core.fields import StreamField
 from wagtail.snippets.edit_handlers import SnippetChooserPanel
-from wagtail.images.edit_handlers import ImageChooserPanel
 
 from taggit.models import TaggedItemBase
 from modelcluster.fields import ParentalKey, ParentalManyToManyField
@@ -51,6 +49,7 @@ base_fields = [
     ('pulse_listing', customblocks.PulseProjectList()),
     ('quote', customblocks.QuoteBlock()),
     ('spacer', customblocks.BootstrapSpacerBlock()),
+    ('looping_video', customblocks.LoopingVideoBlock()),
 ]
 
 
@@ -90,22 +89,6 @@ class BlogPage(FoundationMetadataPageMixin, Page):
 
     zen_nav = True
 
-    hero_image = models.ForeignKey(
-        'wagtailimages.Image',
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name='hero_banner_image',
-        verbose_name='Hero Image',
-        help_text='Image for the blog page hero section.',
-    )
-    hero_video = models.CharField(
-        blank=True,
-        max_length=500,
-        help_text='URL to video for blog page hero section.',
-
-    )
-
     feature_comments = models.BooleanField(
         default=False,
         help_text='Check this box to add a comment section for this blog post.',
@@ -119,13 +102,6 @@ class BlogPage(FoundationMetadataPageMixin, Page):
             heading="Author(s)"
         ),
         FieldPanel('category'),
-        MultiFieldPanel(
-            [
-                FieldPanel("hero_video"),
-                ImageChooserPanel('hero_image'),
-            ],
-            heading="Hero Video/Image"
-        ),
         StreamFieldPanel('body'),
         FieldPanel('feature_comments'),
     ]
@@ -160,12 +136,3 @@ class BlogPage(FoundationMetadataPageMixin, Page):
             context['blog_index'] = blog_page
 
         return set_main_site_nav_information(self, context, 'Homepage')
-
-    def clean(self):
-        if self.hero_image and self.hero_video:
-            raise ValidationError({
-                'hero_image': ValidationError("Please select a video OR an image for the hero section."),
-                'hero_video': ValidationError("Please select a video OR an image for the hero section.")
-                })
-
-        return super().clean()
