+++ network-api/networkapi/wagtailpages/pagemodels/customblocks/__init__.py
@@ -9,6 +9,7 @@ from .image_text_block import ImageTextBlock
 from .image_text_mini import ImageTextMini
 from .latest_profile_list import LatestProfileList
 from .link_button_block import LinkButtonBlock
+from .looping_video_block import LoopingVideoBlock
 from .profile_by_id import ProfileById
 from .profile_directory import ProfileDirectory
 from .pulse_project_list import PulseProjectList
@@ -39,6 +40,7 @@ __all__ = [
     ImageTextMini,
     LatestProfileList,
     LinkButtonBlock,
+    LoopingVideoBlock,
     ProfileById,
     ProfileDirectory,
     PulseProjectList,
