+++ source/sass/main.scss
@@ -41,6 +41,7 @@
 @import "./wagtail/blocks/image-text-mini";
 @import "./wagtail/blocks/iframe-block";
 @import "./wagtail/blocks/link-button.scss";
+@import "./wagtail/blocks/looping-video.scss";
 @import "./wagtail/blocks/rich-text";
 @import "./wagtail/blocks/profiles";
 @import "./wagtail/blocks/article-blocks";
