+++ source/sass/views/blog.scss
@@ -25,8 +25,3 @@ a.category {
 #view-blog-index .result-info {
   font-family: $font-family-slab-serif;
 }
-
-.blog-body {
-  z-index: 3;
-  position: relative;
-}
